backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
pdfminer.six==20231228
pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlite-utils==3.35.2
aiofiles==23.2.1
google-generativeai==0.3.2
pydantic==2.5.0
PyJWT==2.8.0
